requirements.txt
flask
mysql-connector-python
gunicorn
qrcode
Flask-Cors
schedule
requests
pytz
psycopg2-binary